% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{adjusted_ci}
\alias{adjusted_ci}
\title{Adjusted (multivariate) confidence intervals}
\usage{
adjusted_ci(lus, level = 0.95, tol = 0.001, include_unadjusted = TRUE)
}
\arguments{
\item{tol}{tolerance for reaching \code{level}.}

\item{include_unadjusted}{Include unadjusted confidence intervals as an attribute?}
}
\value{
Adjusted confidence intervals with the adjusted confidence level as an attribute.
}
\description{
Adjust coverage level such that joint coverage level corresponds
}
\details{
This function uses output from \code{ciperm0.multi}.
In a simple bisection algorithm, the confidence level is adjusted until the joint coverage level (form \code{alpha.multi}) is \code{code}.
}
