% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{alpha.multi}
\alias{alpha.multi}
\title{Joint coverage level}
\usage{
alpha.multi(lows, ups, alpha)
}
\arguments{
\item{lows}{matrix of l values}

\item{ups}{matrix of u values}

\item{alpha}{Significance level (for unadjusted/single confidence intervals)}
}
\value{
Calculates alpha_multiple from the output of \link{ciperm0.multi}
}
\description{
Joint coverage level
}
