% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cint.R
\name{cint}
\alias{cint}
\title{Permutation-methods confidence interval for difference in means}
\usage{
cint(dset, conf.level = 0.95, tail = c("Two", "Left", "Right"))
}
\arguments{
\item{dset}{The output of \code{\link{dset}}.}

\item{conf.level}{Confidence level (default 0.95 corresponds to 95\% confidence level).}

\item{tail}{Which tail? Either "Left" or "Right" or "Two"-tailed interval.}
}
\value{
Numeric vector with the CI's two endpoints.
}
\description{
Calculate confidence interval for a simple difference in means
from a two-sample permutation or randomization test.
In other words, we set up a permutation or randomization test to evaluate
\eqn{H_0: \mu_A - \mu_B = 0}, then use those same permutations to
construct a CI for the parameter \eqn{\delta = (\mu_A - \mu_B)}.
}
\examples{
x <- c(19, 22, 25, 26)
y <- c(23, 33, 40)
demo <- dset(x, y)
cint(demo, .95, "Two")
}
