% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\docType{package}
\name{ciperm-package}
\alias{ciperm-package}
\title{ciperm: Confidence intervals using permutation tests}
\description{
ciperm implements the functionality of https://arxiv.org/abs/2111.14966.
See more details at also https://github.com/naolsen/ciperm and the  help pages of the individual functions.
}
