% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{ciperm.linreg}
\alias{ciperm.linreg}
\title{Confidence interval for the slope parameter in linear regression}
\usage{
ciperm.linreg(x, y, bounds, ...)
}
\arguments{
\item{x}{Values of the independent variable. A numeric vector of length N.}

\item{y}{Values of the dependent variable. A numeric vector of length N or a matrix of dim N x K.}

\item{bounds}{Bounds forwarded to \code{uniroot} in the search for l and u values.}

\item{...}{Additional arguments to \link{ciperm} or \link{ciperm.multi}}
}
\value{
Vector or array of lower and upper ends of the confidence interval
}
\description{
Calculates the permutation-test based confidence interval for \eqn{\beta_1} in the linear regression model
\deqn{y = \beta_0 + \beta_1 x + \epsilon}. Works in both for univariate and multivariate cases (but \code{x} must be univariate).
}
\examples{
x <- runif(20, -1, 1)
y <- 3 + 0.8*x + rt(20, df = 4)

ciperm.linreg(x, y, bounds = c(-5, 6))


## Multivariate example (requires package \code{mvtnorm})
library(mvtnorm)
x <- runif(20, -1, 1)
sigma <- matrix(0.80, 4, 4)
diag(sigma) <- 1

y <- 3 + 0.8*x + rmvnorm(20, mean = rep(0,4), sigma = sigma)

ciperm.linreg(x, y, bounds = c(-7, 8))

}
\seealso{
\link{ciperm},  \link{ciperm.multi}
}
