% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{ciperm.multi}
\alias{ciperm.multi}
\title{Multivariate confidence intervals using permutation tests}
\usage{
ciperm.multi(
  y,
  phi,
  M = 1000,
  statistic,
  bounds,
  estimate,
  level = 0.95,
  multilevel = TRUE
)
}
\arguments{
\item{level}{Coverage level (i.e. 1- \eqn{\alpha}).}

\item{multilevel}{Calculate the joint coverage level using \link{alpha.multi}?}
}
\value{
A 2 x K array of K confidence intervals.
}
\description{
Constructs confidence intervals at the specified level, and optionally calculates the joint confidence level.
}
\details{
See \link{ciperm0.multi} for additional parameters.
}
\seealso{
\link{ciperm}
}
