% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPerm.R
\docType{package}
\name{CIPerm}
\alias{CIPerm}
\title{CIPerm: Creating CIs for Mean Shift from Permutation Tests}
\description{
Implements computationally-efficient construction of
confidence intervals from permutation tests for simple differences in means.
The method is based on Minh Nguyen's MS thesis paper, 2009,
"Nonparametric Inference using Randomization and Permutation
Reference Distribution and their Monte-Carlo Approximation."
}
