% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{ciperm.twosample}
\alias{ciperm.twosample}
\title{Confidence interval for the two-sample difference in means}
\usage{
ciperm.twosample(x, y, ...)
}
\arguments{
\item{x}{Observations from sample 1}

\item{y}{Observations from sample 2}

\item{...}{Additional arguments to \link{ciperm}}
}
\value{
Vector of lower and upper ends of the confidence interval
}
\description{
Calculates the permutation-test based for the difference in means.
}
\details{
It is the confidence interval for mean(y) - mean(x)
}
\examples{
x <- rnorm(20)
y <- rnorm(15) + 0.3

ciperm.twosample(x,y, M = 2000, level = 0.99) ## 99\% confidence interval, 2000 permutations.

}
\seealso{
\link{ciperm}
}
