% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{ciperm0.multi}
\alias{ciperm0.multi}
\title{Permutation scheme for non-parametric confidence intervals (multivariate version)}
\usage{
ciperm0.multi(y, phi, M = 1000, statistic, bounds, estimate)
}
\arguments{
\item{y}{Matrix of dimension N x K with at least two columns.}

\item{estimate}{Parameter estimate (vector of length K). Optional}
}
\value{
array of dim M x K x 2
}
\description{
The actual permutation scheme, and the multivariate version of \link{ciperm0}.
}
\details{
See \link{ciperm0} for additional parameters.
\code{phi}, \code{statistic} and \code{bounds} are assumed univariate and similar across coordinates
}
