% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{ciperm0}
\alias{ciperm0}
\title{Permutation scheme for non-parametric confidence intervals}
\usage{
ciperm0(y, phi, M = 1000, statistic, bounds, estimate, infinities = FALSE)
}
\arguments{
\item{y}{Data (vector of length N)}

\item{phi}{Covariate function. Must return a vector of length N}

\item{M}{Number of permutations}

\item{statistic}{Test statistic}

\item{bounds}{Bounds forwarded to \code{uniroot} in the search for l and u values.}

\item{estimate}{Parameter estimate. Optional}

\item{infinities}{If TRUE set l_m to infinity, when uniroot fails to find a root  (Not implemented yet).
If FALSE, the procedure fails with an error.}
}
\value{
Matrix of dim M x 2
}
\description{
The actual permutation scheme. Note that \code{estimate} (if provided), \code{statistic} and \code{phi} must match.
}
\seealso{
\link{ciperm}
}
