% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dset.R
\name{dset}
\alias{dset}
\title{dset function}
\usage{
dset(group1, group2)
}
\arguments{
\item{group1}{Vector of values for first group.}

\item{group2}{Vector of values for second group.}
}
\value{
A data frame(?) ready to be used in cint() or pval().
}
\description{
Calculate table of differences in means etc. for each permutation,
as needed in order to compute a p-value and/or confidence interval...
}
\examples{
x <- c(19, 22, 25, 26)
y <- c(23, 33, 40)
demo <- dset(x, y)
knitr::kable(demo, digits = 2)
}
