% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsetFast.R
\name{dsetFast}
\alias{dsetFast}
\title{dsetFast function}
\usage{
dsetFast(group1, group2, nmc = 10000)
}
\arguments{
\item{group1}{Vector of values for first group.}

\item{group2}{Vector of values for second group.}

\item{nmc}{Threshold for whether to use Monte Carlo draws or complete
enumeration. If the number of all possible combinations
\code{choose(n1+n2, n1) <= nmc}, we use complete enumeration.
Otherwise, we take a Monte Carlo sample of \code{nmc} permutations.
You can set \code{nmc = 0} to force complete enumeration regardless of
how many combinations there are.}
}
\value{
A data frame ready to be used in \code{cint()} or \code{pval()}.
}
\description{
DRAFT of faster version of \code{dset()}:
Calculate table of differences in means etc. for each
combination (or permutation if using Monte Carlo approx.),
as needed in order to compute a p-value and/or confidence interval...
}
\examples{
x <- c(19, 22, 25, 26)
y <- c(23, 33, 40)
demo <- dset(x, y)
knitr::kable(demo, digits = 2)
}
