% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{pval function}
\usage{
pval(data, tail = c("Two", "Left", "Right"), value = c("m", "s", "d", "w"))
}
\arguments{
\item{data}{The output of dset().}

\item{tail}{Which tail? Either "Left" or "Right" or "Two"-tailed test.}

\item{value}{Either "m" for difference in means, "s" for sum of Group 1 values,
"d" for difference in medians, or "w" for Wilcoxon rank sum statistic.}
}
\value{
Numeric p-value for the selected type of test.
}
\description{
Calculate p-value from permutation test...
}
\examples{
x <- c(19, 22, 25, 26)
y <- c(23, 33, 40)
demo <- dset(x, y)
pval(demo, "Left", "s")
}
