% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cint.R
\name{cint}
\alias{cint}
\title{cint function}
\usage{
cint(data, sig = 0.05, tail = c("Two", "Left", "Right"))
}
\arguments{
\item{data}{The output of dset().}

\item{sig}{Significance level (default 0.05 corresponds to 95\% confidence level).}

\item{tail}{Which tail? Either "Left" or "Right" or "Two"-tailed interval.}
}
\value{
Numeric vector with the CI's two endpoints.
}
\description{
Calculate confidence interval from permutation test...
}
\examples{
x <- c(19, 22, 25, 26)
y <- c(23, 33, 40)
demo <- dset(x, y)
cint(demo, .05, "Two")
}
