% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{ciperm}
\alias{ciperm}
\title{Confidence interval using permutation test}
\usage{
ciperm(y, phi, M = 1000, statistic, bounds, estimate, level = 0.95)
}
\arguments{
\item{level}{Confidence level (i.e. 1- \alpha).}
}
\value{
Vector of lower and upper ends of the confidence interval
}
\description{
Constructs confidence intervals at the specified level.
}
\details{
See \link{ciperm0} for additional parameters.
}
\seealso{
\link{ciperm.twosample} for a user-friendly version for the two-sample case.
\link{ciperm.linreg} for a user-friendly version for the linear regression case.
}
